# perf: lazily create ureq agents