I actually did a flamegraph on Mac. I didn't notice it was slow because I use Windows, which was way faster at this.

Before:

```
% time dprint check --incremental=false ./data/package-template.json
0.10s user 0.02s system 65% cpu 0.173 total
```

After:

```
% time ../dprint/target/release/dprint check --incremental=false ./data/package-template.json
0.01s user 0.01s system 67% cpu 0.023 total
```